requirements.txt
torch 
torchvision 
torchaudio
llama-cpp-python
pyaudio
wave
openai-whisper
